+++ e8910.h
@@ -3,6 +3,6 @@
 
 void e8910_init_sound ();
 void e8910_done_sound ();
-void e8910_write (int r, int v);
+void e8910_write (uint8_t r, uint8_t v);
 
 #endif
